pyproject.toml
[tool.poetry]
name = "alignment"
version = "0.1.0"
description = ""
authors = ["KarrotLondon <notreal@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
pymongo = {extras = ["srv"], version = "^4.3.2"}
python-dotenv = "^0.21.0"
Flask-Login = "^0.6.2"
pydantic = "^1.10.2"
fastapi = "^0.85.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
